refactor: rename gotest/ -> provisioner/; rework README header + badges
- rename the Go app dir + module to provisioner (github.com/AndriyKalashnykov/
  authentik-k8s/provisioner) — 'gotest' read like the `go test` command
- update all imports, ci.yml (working-directory + paths-filter), CLAUDE.md,
  README, compose/.env.example comment, .gitignore
- README H1 -> repo name + italic tagline (drops doubled 'authentik' + choppy
  parenthetical)
- add missing badges: CI status, Renovate, Go version (provisioner/go.mod)
- binary name (APP_NAME=authentik-poc) unchanged

build + tests pass under the new module path.
