+++ compose/.env.example
@@ -10,7 +10,7 @@ PG_PORT_HTTP=5432
 # --- Authentik ---
 AUTHENTIK_SECRET_KEY=ZnIogyvvZGZRKPNwYjYLNbfJGMyAIi4fYMLvERevXshb81f5X3
 AUTHENTIK_ERROR_REPORTING__ENABLED=true
-# Must match gotest/.env AUTHENTIK_BOOTSTRAP_TOKEN so the Go POC can authenticate.
+# Must match provisioner/.env AUTHENTIK_BOOTSTRAP_TOKEN so the Go POC can authenticate.
 AUTHENTIK_BOOTSTRAP_PASSWORD=Authentik01234567890!
 AUTHENTIK_BOOTSTRAP_TOKEN=NoMlxBQuYgfu3j19ygGqhjXenAjrJgOfN5naqmSDBUhdLjYqHKze7yyzY07H
 AUTHENTIK_IMAGE=ghcr.io/goauthentik/server
