+++ provisioner/.gitignore
@@ -1,3 +1,3 @@
 .idea/
-gotest
+provisioner
 bin/
