+++ provisioner/main.go
@@ -6,7 +6,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/AndriyKalashnykov/authentik-k8s/gotest/internal/authentik"
+	"github.com/AndriyKalashnykov/authentik-k8s/provisioner/internal/authentik"
 	api "goauthentik.io/api/v3"
 )
